only try to link grep if it does not already exist

+++ src/harden/bin/start.sh
@@ -4,7 +4,7 @@ set -e
 echo "---> Starting hardening process"
 
 # cheap hack - nessus for some reason assumes the whole
-ln -s /bin/grep /usr/bin/grep
+[[ -e /usr/bin/grep ]] || ln -s /bin/grep /usr/bin/grep
 
 cd /var/vcap/packages/harden/files
